build: get base of windows build for openzap going again (OPENZAP-55)
git-svn-id: http://svn.openzap.org/svn/openzap/trunk@682 a93c3328-9c30-0410-af19-c9cd2b2d52af

+++ libs/openzap/src/zap_callerid.c
@@ -83,7 +83,7 @@ zap_status_t zap_fsk_data_add_checksum(zap_fsk_data_state_t *state)
 		check = check + state->buf[i];
 	}
 
-	state->checksum = state->buf[state->bpos] = 256 - check;
+	state->checksum = state->buf[state->bpos] = (uint8_t)(256 - check);
 	state->bpos++;
 
 	state->dlen = state->bpos;
